requirements.txt
# Core Django & PostgreSQL
Django==4.2
psycopg2-binary==2.9.9
python-dotenv

# REST API & Django Extensions
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
drf-yasg==1.21.7  # Swagger for API docs

pytz
solana
base58

# Web3 & Blockchain Interaction (Solana)
# solana==0.34.0
# pyserum==0.5.1  # For interacting with Serum DEX on Solana
# web3==6.10.0  # If you plan to extend for Ethereum later

# Data Handling & Trading Analysis
# pandas==2.1.3
# requests==2.31.0

# # Task Scheduling & Async Execution
# celery==5.3.4
# redis==5.0.1  # For Celery task queue
# django-celery-beat==2.5.0  # Periodic tasks